pyproject.toml
[project]
name = "push-notification"
version = "0.1.0"
description = "Daily push notification to Telegram via GitHub Actions."
readme = "readme.md"
requires-python = ">=3.13"
dependencies = [
  "beautifulsoup4>=4.14.3",
  "lxml>=6.0.2",
  "playwright>=1.57.0",
  "pyyaml>=6.0.3",
  "requests>=2.32.0",
]

[project.scripts]
push-notify = "push_notification.main:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"
